+++ gridsome/lib/app/ContentTypeCollection.js
@@ -91,6 +91,9 @@ class ContentTypeCollection extends EventEmitter {
     node.title = options.title || fields.title || node.title
     node.date = options.date || fields.date || node.date
     node.slug = options.slug || fields.slug || this.slugify(node.title)
+    node.content = options.content || fields.content || node.content
+    node.excerpt = options.excerpt || fields.excerpt || node.excerpt
+    node.internal = Object.assign({}, node.internal, internal)
     node.path = typeof options.path === 'string'
       ? '/' + options.path.replace(/^\/+/g, '')
       : this.makePath(node)
@@ -259,7 +262,7 @@ class ContentTypeCollection extends EventEmitter {
     // with '_raw' suffix.
     for (let i = 0, l = routeKeys.length; i < l; i++) {
       const keyName = routeKeys[i]
-      const fieldValue = node.fields[keyName] || keyName
+      const fieldValue = node.fields[keyName] || node[keyName] || keyName
 
       if (
         isObject(fieldValue) &&